[velox] refactor(functions): Move Spark decimal ceil struct to header
GPU SFI needs to compile Spark decimal ceil `call()` bodies from device TUs. `DecimalCeilFunction` lived in an anonymous namespace in `DecimalCeil.cpp`, so no other translation unit could `#include` it.

Move the struct verbatim into `DecimalCeilFunction.h` under `facebook::velox::functions::sparksql`. Keep `registerDecimalCeil()` in `DecimalCeil.cpp`. No logic or registration API changes.

Issue: https://github.com/facebookincubator/velox/issues/17259

Test plan:
- Build and run `velox_functions_spark_test --gtest_filter=DecimalCeilFloorTest.*` in the container
